Add support for pty and cgroup pseudo-filesystems.
The boot utility was modified by this update to mount the
pseudo-filesystem which is mounted at /dev/pts to support the
creation of pseudo-ttys.  In addition the cgroup
pseudo-filesystem is mounted on the /sys/fs/cgroup mountpoint
prior to the switch to the root filesystem.

All of this functionality is implemented in order to support the
ability of the security supervisor to use runc to execute
canisters.

This will require that cgroups be enabled in the operating system
image or the switch to the root filesystem will fail.