Add initial implementation of cboot utility.
This update implements a utility which will ultimately be used to
run a behavioral canister which is capable of propagating its
information exchange events into an SGX container or some other
secured framework.

The behavior namespaces expose their measurement events through
the following pseudo-file:

/sys/fs/iso-identity/update-NNNN

Where NNNN is replaced with the namespace inode number which is
allocated when the behavioral namespace is completed.

The cboot utility polls this pseudo-file and when data becomes
available reads the measurement event from the pseudo-file.
Currently it simply prints out the event but the initial
implementation of this utility will propagate the measurements
into an SGX enclave.