Implement test harness for SGX based measurement enclave.
This update introduces an SGX enclave which is used to implement
SHA256 based measurement registers.  This enclave is an initial
proof of concept implementation which is expected to be enhanced
to support multiple registers.

The enclave currently supports the ability to accept an arbitrary
buffer and extended the current measurement state with the
measurement of that buffer.  The current implementation arguably
needs to be modified to first reduce the buffer to a 256 bit
measurement and use that binary value to extend the current 256
bit measurement.

The enclave also supports the ability to read the current
measurement value from the enclave.

This update also introduces the test-pcr utility which exercises
the functionality of the enclave.  The utility currently executes
a loop which requests input from the user and passes the string
typed by the user into the enclave to be used as the measurement
extension.