Correctly handle build of sub-directories.
The lib directory which builds the NAAAIM support library was
integrated into the SUBDIRS variable which caused builds in the
root directory of the source tree to fail.  Handling of the lib
directory was broken out separately and is now conducted first in
order to allow the root directory and all of the subordinate
subdirectories to build correctly.

A dependency clause for the SGX sub-directory was missing which
caused that directory not to be built when the build was invoked
from the root of the source tree.  This update adds the proper
dependency clause.

Both builds and cleans of the source tree now work correctly.