Drop sgx-test utility.
The sgx-test utility was originally written as part of the
initiative to begin developing an independent implementation of
the Linux SGX loader.  Retaining it makes little sense at this
time since richer functionality is available in the loader
utilities.

In addition the changes in the kernel driver API which will be
the subject of the next committment make maintaining this utility
problematic.