Update to new kernel driver API.
The kernel driver API was modified after the 1.6 release of the
SGX Software Development Kit to move the creation of the enclave
memory map from the kernel to userspace.  This change in the
driver was incompatible with the initial implementation of the
SGX loader which targeted the former API.

This update modifies the SGX loader environment to use the new
API.  This involves issueing an mmap() call as part of the
enclave creation process to create a memory region which meets
the SGX alighnment requirements.

The kernel API change drops the variant of the ioctl which
destroys the enclave.  This is no longer needed since executing a
munmap() call against the enclave pointer address has the same
effect.  The SGXenclave destructor replaces the ioctl call with
an appropriate munmap call to release the enclave mapping.