Update SGX PCR measurement utility.
The enclave which implements the emulation of a Platform
Configuration Register was modified to compute the measurement
based on computing the hash over the concantenation of two
256-bit numbers.  The computed measurement now matches how PCR
registers are handled inside of a TPM.

This change involved computing the SHA256 hash value of the input
arguement and using that instead of the buffer which was input to
the extend ECALL.

The function which loops over user input passing the string which
was input to the extend ECALL now removes the linefeed from the
input string since it is technically not relevant to the
computation of the measurement value.  In addition the input is
checked for the string 'quit' which is used to signal the end of
input.

The idf-sgx-fandf-unibin was updated to use the sgx-pcr.signed.so
enclave so PCR measurement could be used to demonstrate the
unified binary support.