Implementation of a UNIX domain management object.
This update implements a LocalDuct object which is used to manage
communications through a UNIX domain socket.  From a workability
perspective it is designed to be similar in feature and
capabilities to the Duct object which implements similar
functionality over the network.

The LocalDuct_test utility is the unit test for this object.