Add methods to implement access to socket descriptors.
The ->get_socket and ->get_fd methods were added which implement
the ability to retrieve the file descriptors which reference the
socket itself and the accepted connection.

These are needed in order to allow the ability to select and/or
poll for new socket connections and the presence of data
availability on a connection.