Fix socket close behavior.
If a communications socket is open the previous version of this
code would only close the descriptor if this the LocalDuct object
was a server instance.  This was changed to close the descriptor
regardless of the instance type of this object.

The previous version of this code would unconditionally unlink
the filesystem instance of the socket.  This update changes the
behavior to only unlink the socket if object is a server
instance.  This allows clients to come and go without rendering
the the server process unreachable due to the absence of a
filesystem socket instance.