Major functionality changes to canister boot system.
This update implements a number of major improvements to the
cboot utility which is used to boot and monitor a canister
instance.

The most significant functional change is the implementation of a
management utility which connects to the cboot daemon and allows
commands to be issued to the canister instance.  Other then a
directive to terminate the management utility instance the only
command implemented is the ability to return the current
measurement state of the canister.

The cboot utility itself was modified to take the ASCII
hexadecimal measurement value from the /sys pseudo-file which
surfaces the measurements from the kernel and uses that value to
extend the current measurement state which is maintained for this
canister.