Implement signal handler for cboot daemon.
Up to this point issueing a CNTRL-C key combination to the
canister boot utility caused the utility to immediately halt
without calling the cleanup routines.  As a result the filesystem
socket instance was left which required manual cleanup.

Signal handling was added to catch and react to the SIGINT
signal.  On an error from the poll() call the signal status
structure is queried and if the SIGINT signal was the cause for
the error a normal program exit is taken.  This causes the object
destructors to be called which result in removal of the socket
instance.