Add SGX measurement enclave support.
This update implements the ability for the cboot utility to use
an SGX enclave implement canister measurement support.  The
ability to implement canister measurement within the cboot
utility itself is retained.

SGX enclave support is enabled with the -S switch to cboot.  A
processor specific EINITTOKEN must be supplied with the -t
command-line switch if enclave mode is selected.

Upon request from the cboot-mgr the canister measurement is
returned either from the enclave or the value maintained
internally in the utility itself.

The enclave expects the measurement value to be submitted to the
enclave in binary form.  The enclave hashes this to a 256-bit
measurement value and then uses this hash to extend the current
measurement value maintained in the enclave.