Modify signal handling.
This update adds trapping of the SIGTERM, SIGQUIT and SIGHUP
signals.  Rather then setting an individual flag for each signal
type the signal handling code was modified so that all three of
these signals plus SIGINT set a common flag indicating the
application should stop.

This is still insufficient to properly capture the termination of
the program when the initial process of a container is exited.