Cleanup unreleased objects.
The SGXenclave object was not being destroyed on program exit.  A
destructor for this statically declared object was added to the
end of program cleanup processes.

In the case of SGX mode the buffer which was used to hold the
incoming mesurement was not being released.  The if clause which
codes for SGX mode processing was changed to set a return value
and use the standard exit processing which does include a
destructor for the object.