Invoke canister execution with runc utility.
This commit implements the ability for cboot to start a canister
environment using the OCI compliant runc utility.

In order to support this model the cboot utility unshares its
behavioral namespace after setting up its management socket and
optionally the SGX measurement enclave.  The utility then forks
and executes the runc utility from the user's path definition.

The cboot utility was enhanced to now accept the -n and -b
command-line switches.  The arguements to these are the name of
the canister and an option 'bundle' directory which specifies the
root of the container.

The signal_action function was modified to check for the presence
of the SIGCHLD signal on invociation.  If set the sigchild member
of the Signals structure is set to true.  The error handling
clause of the poll utility was modified to check for the presence
of SIGCHLD and if found the cboot utility is terminated.

After unsharing its behavioral namespace the cboot utility
queries the /proc/self/ns/behavior pseudo-file to obtain the
inode number of the namespace.  This is used to construct the
path to the /sys/fs/iso-identity/update-N pseudo-file which the
cboot utility will poll for behavior events.