Implement generic event processing.
The initial implementation of userspace processing of
iso-identity information exchange events focused only on emitting
the measurement value which would be normally used to extend a
hardware TPM.

In order to support a generic userspace implementation of
iso-identity modeling the cboot event processing code now interprets
the first space delimited keyword of the event as the type of
event represented by the remainder of the string.  This keyword
is used to vector processing of the remainder of the string.