Implement initial support for information exchange events.
The add_event() function was added to implement the processing of
an information exchange event exported from the kernel.  An
information exchange event is the most fundamental interaction in
iso-identity modeling and describes the interaction of an actor
and subject identity.

The parsing of the event information is implementing by using the
Actor and Subject objects which were originally created to
process trajectory maps.  The information exchange event is the
same information which is exported by each line of the
/sys/kernel/security/ima/iso-identity/trajectory pseudo-file.

The current implementation only prints out the event name and the
measurement value of the Actor and Subject identities.