Implement initial information exchange event processing.
This update adds the add_event() function to process information
exchange events.  This function adds the current event to the log
of events maintained by this instance of the container daemon.

The cboot-mgr utility and the canister management daemon was modified
to process the 'show trajectory' command which requests that the event
log be output to the management instance.

The input processing loop in the boot utility was modified to
simplify its implementation.  The loop now appears to properly
exhaust the contents of the /sys/fs/iso-identity/update-NN
pseudo-file so that at any time the canister management daemon
has the same model of the namespace that the kernel does.