Add object for managing information exchange events.
The ExchangeEvent object was added in this update.  This object
implements various methods for processing and measuring an
iso-identity information exchange event.  It operates at this
revision level as a wrapper for the Actor and Subject objects and
provides a one step higher level of abstraction over these
objects.

The generate-contours utility was updated to the ExchangeEvent
object to compute the contours for a given trajectory.  The code
which implemented a measurement value over the intersection of an
Actor and Subject identity was moved into the ->measurement
method of the ExchangeEvent object.

The updated binary was verified so that it produces a set of
contours consistent with those produced by a canister over the
same trajectory path.