Add missing numeric object identifier.
The update which added the ExchangeEvent object failed to include
the change to NAAAIM.h object identification which added the
numeric object identifier for the object.  This commit adds that
identifier.

The previous commit this will not compile but the fix is
reasonably trivial and unlikely to cause issues for any possible
bisections.