Use canister name to compose management socket name.
The UNIX domain management socket had previously used the
following format for the socket name:

cboot-mgmt.NN

Where NN was replaced withe proced id (PID) of the instance of
the canister daemon.  This was somewhat cumbersome and was thus
replaced with the following format for the name:

cboot-mgmt.NAME

Where NAME is the canister instance name which is specified with
the -n switch to the cboot daemon.

The cboot-mgr daemon was modified to change the -f switch which
specified the path to the management socket with the -n switch
which specifies the name of the canister instance which is to
be managed.