Add ISOidentity modeling object.
This commit introduces an object named ISOidentity which
implements the iso-identity modeling system in userspace.  It is
designed to implement the same functionality as is available
through the kernel implementation but through a userspace
implementation.

The test-ISOidentity program was added to implement a test
harness for testing behavioral trajectories using this modeling
object.