Implement a utility to compute aggregate measurement values.
This update implements a simple utility program named
compute-aggregate which can be used to generate the SHA256
aggregate value which is generated by the kernel IMA
implementation.  The utility takes two arguements which are the
actor and subject identities from the first line in the following
pseudo-file:

/sys/kernel/security/ima/ascii_runtime_measurements

The 256 bit value generated will be the measurement which is at
the root of the measurement list either for the kernel
implementation or a userspace simulation of the kernel
implementation.