Convert cboot to use ISOidentity modeling object.
The cboot utility was updated to use the ISOidentity modeling
object to implement the modeling of the canister which is being
disciplined.  Since the modeling object properly tracks duplicate
trajectory points the trajectory provided by the cboot utility
now matches the trajectory which is available inside the canister
itself.

The measurement values are currently different since the
ISOidentity object uses a default null aggregate value for its
measurement.  The canister should output the currently active
aggregate value in order for the measurements to properly match.

An independent trajectory is maintained in the cboot utility in
the form of the event string.  The utility needs to be updated to
access the trajectory which is already available inside of the
ISOidentity modeling object since this is a duplicate of the
information which is available.