Implement support for capturing event definition.
The ExhangeEvent object was modified to parse the event
description field and to store that field as a String object in
the state information of the exchange event.

The event description field is as follows in the trajectory:

	event{COMM:PATH}

The COMM field is the name of the process involved in the event
while the PATH is the pathname of the subject.

The ISOidentity object was modified to describe each point in the
following form:

	Point NN:

Where NN is the sequence number in the trajectory path.