Add support for process identifiers in exchange events.
The ExchangeEvent object was modified to add support for parsing
of an option process identifer field in an ASCII encoded
information exchange event.  The format of the field is as
follows:

	pid{NNN}

Where NNN is the process identifier of the information exchange
event.  If an process identifier has not been specified the pid
defaults to a value of 0 which is not a possible value for a
kernel generated identifier.