Add support for accessing process identifier.
The ->get_pid method was added for accessing the process
identifier value associated with an information exchange event.

The cboot utility was modified to demonstrate the use of this
functionality.  The primary use will be when support is added to
the iso-identity modeling system in the kernel to set the actor
status of a process.