Implement retrieval of trajectory events.
The ISOidentity object was updated to implement retrieval of the
trajectory map.  This involved addition of the ->get_event and
->rewind_event methods.

The ->get_event method is designed to be called iteratively until
all points in the trajectory path have been traversed.  This is
signified by the returned event being NULL.

The ->rewind_event method resets the trajectory path to its
starting point.

The ->dump_events method was modified to use this new
infrastructure.