Implement retrieval of behavior contour points.
The ISOidentity object was updated to implement retrieval of the
behavioral contour points.  This involved addition of the
->get_contour and ->rewind_contours methods.

The ->get_contour method is designed to be called iteratively until
all points in the behavioral map have been traversed.  This is
signified by the returned contour being NULL.

The ->rewind_contours method resets the behavioral map to its
starting point.

The ->dump_contours method was modified to use this new
infrastructure.