Implement generation of Subject identity elements.
The ->format method was added to the Subject object to generate an ASCII
formatted set of identity elements.  This formatted string is in the
format which is expected by the ->parse method.

The test-subject utility was updated to exercise this method by
outputting the subject clause of the sample event which was parsed.