Added generation of formatted exchange events.
This update adds the ->format method to the ExchangeEvent object
building on similar support which had been added for the Actor
and Subject objects.  The ExchangeEvent ->format method will now
generate an ASCII formatted version of the information exchange
event which is generated by the in-kernel modeling engine.

The process of adding this method led to the need to make
functional modifications to the Actor and Subject ->format
methods.

The strategy of using snprintf to write directly into the String
object in the Actor->format method turned out to be too fragile.
The method was modified to print into an on-stack buffer which
was then added to the String object supplied as an arguement to
this function.

The Subject->format method was appending a newline to the string
which turned out to be incorrect as these methods are expected to
output standard null-terminated strings.