Add test program for ExchangeEvent object.
The test-event utility was added as a unit test for the
ExchangeEvent object.  It parses an internally supplied
information exchange event screen and outputs the component
identity information about the event.

The utility regenerates the event string with the
ExchangeEvent->format method and then re-parses that string to
verify that the measurement of the event is consistent with that
produced by the test string.  This validates the fidelity of the
text parsing and inversion process.