Implement seeding iso-identity model with aggregate measurement.
Support was added to the cboot daemon for the following event:

	aggregate SHA256

Where SHA256 is the hardware aggregate measurement which is the
first measurement committed to the event log.

Integrating the hardware aggregate value into the canister
measurement ties the canister to the machine which the canister
is running on.

With this support and corresponding support in the kernel the
measurement of a canister is now consistent with the measurement
value maintained by the in-kernel modeling engine.