Implement support for sealing an ISOidentity domain.
This update adds the following methods to the ISOidentity object:

	->seal
	->get_forensics
	->rewind_forensics
	->forensics_size
	->dump_forensics

The ->seal method instructs the identity model to consider any
additional updates which are not in the current behavioral map to
be forensics violations.  The ->dump_forenscis method is a
diagnostic method which outputs the events which have been
captured as forensics violations.

The ->get_forenscis, ->rewind_forensics and ->forensics_size
methods are used to traverse the forensics event list.

The test-ISOidentity utility was updated to allow an artifical
forensics event to be injected into it for test purposes.