Implement support for canister sealing.
The cboot daemon was modified to handle the following event directive
from the kernel:

	seal

Which the cboot utility uses to seal the ISOidentity model object
being maintained for the canister.

The cboot-mgr utility was updated to allow it to send the
following commands to the management daemon:

	show forensics

The canister modeling daemon responds to this command by sending
the list of forensics events which have been registered for this
canister.