Implement ContourPoint to abstract behavior map points.
This update converts the ISOidentity model object to use a
ContourPoint object to represent points in the behavioral map.

In addition to the binary measurement point the object
incorporates a representation of whether or not the point is
considered a valid or invalid measurement point.  This will be
needed to support the disciplining of processes which are
determined to have engaged in extra-dimensional behaviors.