Add support for disciplining a process.
This commit provides an initial framework for using the
set_bad_actor system call to interrogate and/or set the bad actor
status of a process.

Bad actor status means a process has engaged in an
extra-dimensional information exchange event after the platform
behavior model has been sealed.  This status bit is queried by
the Autonomous Introspection (AI) Security module to determine
actions which the process is allowed to implement.

The set_bad_actor system call takes a process identifier (PID) as
an arguement along with a flags arguement.  A flag arguement of 0
indicates the status of the specified process should be queried.
A flag arguement of 1 indicates the specified process is to have
its actor process set.

The current disciplining code checks the 'bad actor' status of a
process when an extra-dimensional event is detected.  If the
status is not set the utility issues the system call to set the
status.  While not incorporated it the management daemon can
optionally kill the offending process if it is not response to
the setting of its bad actor status.