Initial implementation of the bad-actor utility.
This update provides an initial implementation of a utility which
allows the 'bad actor' status of a process to be queried and/or
set.  From a query perspective this information is also available
in the following pseudo-file:

/proc/PID/status

As the 'bad actor:' tag.

This utility also provides the ability to set the 'bad actor'
status flag which is an irrevocable operation.