Initial implementation of SGX HurdLib.
The HurdLib library is the foundational library which all of the
IDfusion infrastructure is built on.  This commit implements an
initial implementation of the library which is capable of running
in an enclave.

The NAAAIM library will be implemented next on top of this
library.

The strategy will be to use these two libraries to provide an
abstraction layer so that only a single instance of, for example,
the ISOidentity modeling code can be created which will run,
largely without changes, in both a standard userspace as well as
an enclave.

The SGXfusion library is not yet linked into the general build
structure.  That will be implemented after the framework for the
NAAAIM library is established.

Both libraries will not be exact duplicates of their untrusted
userspace versions.  Both libraries will have SGX specific code
in their build infrastructure in order to implement needed
functionality and transparency for the objects which are common
between both the trusted and untrusted versions.