Initial NAAAIM library implementation.
This update provides the framework for building an NAAAIM object
library which is capable of running in an SGX enclave.  This
library builds on top of the SGXfusion/HurdLib library which was
previously committed.

The initial object implemented is the SHA256 object which
provides support for generating SHA256 based checksums on
arbitray input buffers.