Force inclusion of SGXfusion library.
An issue was noted with enclave initialization failing.  The
EINIT ioctl was failing with a value of 4 which indicates that
enclave measurement is failing.

After an extensive and lengthy round of debugging this could
be reproduced by an enclave containing only a single ECALL
containing only a print statement or a simple return statement.

The root cause of this was found to be the need to reference an
object in the SGXfusion library.  Referencing, for example, the
Buffer object would cause the enclave to initialize properly.

This was fixed by including the SGXfusion library in the
--whole-archive clause.  This causes the entire contents of the
library to be linked into the shared image.  It is currently
unclear why the shared image would successfully link but then
fail enclave measurement at the time of initialization.

This update fixes the Makefiles for both the SGXfusion and
SGXnaaaim libraries.  It also corrects and standardizes the
formatting of these files.