Modify loading of unitialized data.
Work on the creation of enclaves which supported regular
expression processing using code imported from the MUSL library
caused enclaves to fail EINIT processing with an error code of 4
which indicates an incorrect enclave measurement.

Significant effort was expended on isolating the source of the
regression which turned out to be how the loader was handling
unitialized data from the segment which is composed of Block
Static Storage (BSS) data.  Adding a 'dummy' char array which was
varied in size caused the measurement error to be corrected
and/or to re-appear.

The sentinel finding appears to be that the error is provoked by
a non-zero offset value for the data being loaded.

The following commit in the Intel reference implementation:

14c19edaafdd
"Clean up build_sections and build_mem_region code

Modifies the behavior of the equivalent section of code in the
reference implementation.

I had to resort to a heuristic to get the SGXloader object to
properly load unitialized data and even with this heuristic it
appaers there may be a problem with loading.  I have concerns
that this may be secondary to the use of the Intel sgx_sign
binary which is using the older variant of their loader.

If my concerns are legitimate Intel may have an issue where an
enclave which is signed with an older version of the loader code
may have issues being loader by a newer variant of the code.