Provide build linkage to ISOidentity modeling objects.
The rationale of the SGXfusion and NAAAIM libraries is to allow
enclave based variants of userspace code to be compiled without
changes to the code itself.  This update implements the ability
to build the userspace ISOidentity objects by providing a generic
build role which allows objects to be compiled from the source
code found in the userspace variant of the modeling engine.

The ContourPoint object was added as the first test of this build
support.