Implement enclave support for information exchange events.
This update links in compilation of the following ISOidentity
modeling objects:

ContourPoint

Actor

Subject

ExchangeEvent

The ISOidentity enclave was modified to process each update event
through the ExchangeEvent object.  The state of the object is
dumped out as verification of proper processing of the object.

A type definition was created in the SGXfusion header file
infrastructure since the Intel standard C library has no notion
of processes.