Explicitly set enclave internal argument to NULL.
The enclave_update variable which is used to hold the enclave
internal string arguement could have been used unitialized in the
event of the ECALL being given a NULL arguement.  Explicitly
initialize the pointer to a null value to prevent a random value
from being passed to the free() call.