Provide Buffer interface to ContourPoint measurement value.
A regression was noted in the ability of the ISOidentity to dump
out the current contour map.  This was found to be secondary to
the fact that ContourPoint objects were being placed on the
contour list rather then Buffer objects which had previously been
used.

The ->get_Buffer method was added to the ContourPoint object
which allows the measurement value to be retrieved as a Buffer
object.  The ->dump_contours method uses this to retrieve the
contour point value for printing.

This interface change now requires that the ContourPoint header
definition be included as part of any file which implements the
ISOidentity model.