Implement enclave based ISOidentity modelling.
The update integrates the ISOidentity modelling object into the
build and updates the enclave to implement this model in an
enclave.  This completes the process of having the ability to do
ISOidentity modelling within the protection of an enclave.

In addition to the update ECALL this update provides the
following three additional ECALL's:

init_model

seal_model

dump_model

The init_model call must be made before any additional ECALL's
are made to the enclave.  This call initializes and creates the
model object which subsequent transactions will execute against.

This update also provides the framework for making the
development of SGX interfaces more straight forward.  The
following defines are now used to configure the ECALL and OCALL
tables:

ECALL_NUMBER

OCALL_NUMBER

Which are used to define the number of ECALL's and OCALL's which
will be supported for the enclave.