Implement model size interface to ISOidentity modelling enclave.
The update provides an ECALL interface to return sizes of various
aspects of the model being run inside of the enclave.

In order to reduce the total number of ECALL's which need to be
defined a multiplexing interface was implemented.  An integer
arguement to the call specifies which model size is to be
returned.

The following encoding values are used:

	0: The size of the behavioral topology map.

	1: The size of the forensics topology map.

The default return value is 0 which may also imply the model is
not initialized.