Isolate enclave interfaces into separate header file.
This update begins implementation of a strategy where the
interfaces to the enclave are isolated into a separate header
file which can be used in both the untrusted code which is
calling into the enclave as well as the enclave interface
functions.

This required placing the external function prototypes into the
file which defines the interface functions.  If enclave code is
split into multiple independent compilation units the strategy
will be to have a separate include file for each compilation unit
listing its external definitions.

Using this strategy allows the interface structures to be defined
only use.  It also provides a method important in this case for
defining constants which can be used to select types of
functionality inside of the enclave.