Implement setting of aggregate measurement value.
This commit implements an ECALL interface for setting the
aggregate hardware value which the ISOidentity model will operate
under.

The ECALL provides a method for conveying the aggregate value in
binary form to the enclave so that it can be introduced into the
module with the ISOidentity->set_aggregate method.