Add support for retrieving disciplined PID from ISOidentity enclave.
An ECALL was implemented which provided the ability to obtain the
process ID of the event which had resulted in an
extra-dimensional behavior event.  This ECALL provides an
interface to the following ISOidentity method:

->discipline_pid

The interface to the update_model ECALL was modified to include a
boolean value which is used to signal back to the caller that
a disciplining event has occurred.

The test harness was modified to include a test of this value
after a forensic event was injected into the model.  The sample
input was modified to include a pid{NN} description field for the
forensics event.  The test harness prints out the value of the
PID to confirm that an event is being signalled and that action
can be taken on this signal.