Create ECALL interface to rewind model components.
This commit provides infrastructure to support the reset of one
of the model components.  This call provides an interface to the
following ISOidentity methods:

->rewind_event

->rewind_forensics

->rewind_contours

A single ECALL is used with a type arguement which selects the
model component to be reset.  This strategy is used to reduce the
amount of code needed to implement the needed interfaces.

The test harness will be updated to extract forensics and
standard model events in the next update and this interface will
be tested.