Implement retrieval of model events.
The commit implements an ECALL which allows formatted model
update events to be returned from the the model being implemented
in the enclave.

A regression was noted in the previous commit which was caused by
the wrong enclave function being called from the interface code.
This was preventing the the rewind requests from being carried
out.

The test harness was updated to playback both the model update
and forensics events using this new infrastructure.