Set return value for get_event ECALL.
While working on the ISOenclave object an error was found while
attempting to use the get_event ECALL.  This was soruced to the
fact the return value was not being explicitly for a successful
return of an event.

A return value of true was explicitly set for the successful
return of an event which addressed the regression.