Implement ISOenclave object.
Work in developing the test-ISOidentity-enclave test harness
indicated that it would be advantageous to encapsulate all of the
infrastructure needed to call into the ISOidentity modelling
enclave.  The ISOenclave object was created to provide an object
based interface into the modelling enclave.

The ISOenclave object was created largely by creating a copy of
the ISOidentity object itself.  Due to the fact that the ECALL
based interface into the modelling enclave is slightly different
then the native ISOidentity model API there are some method
changes which were implemented.

Most notably the routines which return either standard or
forensic events return a string containing the event itself
rather then an ExchangeEvent object.