Fix ECALL slot number for set_aggregate function.
It was noted in testing that the measurement of the model was
different for the model being run in an enclave then the value
that was generated for in-kernel and non-enclave ISOidentity
implementations.  The latter two were generating measurement
values which were identical but different.

This was traced to the fact that due to cut and paste of code the
ECALL slot number for the call in the set_aggregate() function
was not correct.  A value of zero was being specified rather then
the correct value of 5.  This was corrected and all three model
implementations now return the correct value.

This highlights one of the issues which needs to carefully
observed with this ECALL API slot definition architecture.
Consideration needs to be given to integrating the slot number
into the arguement marshalling structure so that input to the
call can be verified as being compatible.