Implement enclave modelling.
This update integrates use of the ISOenclave object which has
been the subject of recent development into the cboot utility.
Use of this object replaces the preliminary and prototype enclave
implementation which was previously in the code for proof of
concept purposes.

The utility was tested and the use of standard userspace
modelling or enclave based measurements can be selected on the
command-line.  This produces a fully functional demonstration of
the use of enclaves to protect iso-identity OS modelling.