Implement support for target and report data structures.
The SGX header definitions were updated to include definitions
for the SGX_targetinfo and SGX_report structures.  The former
structure is used as input the ENCLU EREPORT instruction.  The
latter structure is the output of the instruction.

The SGXenclave object was updated to add the ->get_targetinfo
method which populates a SGX_targetinfo.  The method is only
valid after the enclave has been opened and loaded.

The sgx-load test harness was updated to exercise this new method
and output the elements of the structure.