Remove unneeded ioctl include and SGX definition.
As the code has matured the include of the linux/ioctl.h is not
needed.  This can be addressed by including this header file in
advance of any file which needs the SGX.h header.

Also the SGX_DEVICE define is not useful as the SGXenclave method
call explicitly specifies the device node to be used.