Implement initial support for report generation.
This update adds the Attestation directory immediately under the
SGX directory.  The former directory will be used to hold
infrastructure which implements both intra-machine enclave
verification as well as inter-machine enclave verification.

The first step in this process is to demonstrate the ability to
execute the ENCLU[EREPORT] instruction inside of an enclave.  A
minimal enclave which simply executes and returns the report is
implemented in the form of the LocalTarget.so enclave.

The enclu_ereport function was implemented in assembler code as
part of the SGXfusion library.

The test-local-target.c test utility was implemented to request
generation of a report and the output of the fields in the report
data structure.